pyproject.toml
[project]
name = "vps-config"
version = "0.1.0"
description = "Personal VPS configuration with Ansible"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ansible>=6.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "requests>=2.28.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "remnawave>=2.6.3",
]

[project.scripts]
vps = "vps_cli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["src/vps_cli"]
